move utilities and patch Lightning dataloader methods
