+++ pytorch_lightning/lite/lite.py
@@ -29,13 +29,17 @@ from pytorch_lightning.lite.wrappers import (
     _LiteDataLoader,
     _LiteModule,
     _LiteOptimizer,
-    _replace_dataloader_init_method,
 )
 from pytorch_lightning.plugins import DDPSpawnPlugin, DeepSpeedPlugin, PLUGIN_INPUT, TPUSpawnPlugin, TrainingTypePlugin
 from pytorch_lightning.trainer.connectors.accelerator_connector import AcceleratorConnector
 from pytorch_lightning.utilities import _StrategyType, DeviceType, move_data_to_device
 from pytorch_lightning.utilities.apply_func import apply_to_collection, convert_to_tensors
-from pytorch_lightning.utilities.data import _auto_add_worker_init_fn, _update_dataloader, has_iterable_dataset
+from pytorch_lightning.utilities.data import (
+    _auto_add_worker_init_fn,
+    _update_dataloader,
+    has_iterable_dataset,
+    _replace_dataloader_init_method,
+)
 from pytorch_lightning.utilities.device_parser import _parse_devices
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.seed import seed_everything
