+++ pytorch_lightning/lite/wrappers.py
@@ -11,11 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import functools
-import inspect
-from contextlib import contextmanager
-from itertools import chain
-from typing import Any, Callable, Generator, Iterator, Optional, Set, Type, Union
+from typing import Any, Callable, Generator, Iterator, Optional, Union
 
 import torch
 from torch import nn as nn
@@ -108,49 +104,6 @@ class _LiteModule(DeviceDtypeModuleMixin):
         to_type = torch.get_default_dtype()
         output = apply_to_collection(output, function=_convert_float_tensor, dtype=Tensor)
         return output
-
-
-def _wrap_init(init: Callable) -> Callable:
-    """Wraps the ``__init__`` method of the dataloader in order to enable re-instantiation of custom subclasses of
-    :class:`~torch.utils.data.DataLoader`."""
-
-    @functools.wraps(init)
-    def wrapper(obj: DataLoader, *args: Any, **kwargs: Any) -> None:
-        params = dict(inspect.signature(obj.__init__).parameters)
-        params.pop("args", None)
-        params.pop("kwargs", None)
-        for arg_name, arg_value in chain(zip(params, args), kwargs.items()):
-            setattr(obj, arg_name, arg_value)
-        init(obj, *args, **kwargs)
-
-    return wrapper
-
-
-# https://stackoverflow.com/a/63851681/9201239
-def _get_all_subclasses(cls: Type[Any]) -> Set[Type[Any]]:
-    """Returns a list of all classes that inherit directly or indirectly from the given class."""
-    subclasses = set()
-
-    def recurse(cl: Type[Any]) -> None:
-        for subclass in cl.__subclasses__():
-            subclasses.add(subclass)
-            recurse(subclass)
-
-    recurse(cls)
-    return subclasses
-
-
-@contextmanager
-def _replace_dataloader_init_method() -> Generator[None, None, None]:
-    """This context manager is used to add support for re-instantiation of custom (subclasses) of
-    :class:`~torch.utils.data.DataLoader`. It patches the ``__init__`` method."""
-    for subclass in _get_all_subclasses(DataLoader):
-        subclass._old_init = subclass.__init__
-        subclass.__init__ = _wrap_init(subclass.__init__)
-    yield
-    for subclass in _get_all_subclasses(DataLoader):
-        subclass.__init__ = subclass._old_init
-        del subclass._old_init
 
 
 class _LiteDataLoader:
