+++ pytorch_lightning/trainer/connectors/data_connector.py
@@ -19,6 +19,7 @@ from weakref import proxy
 
 import pytorch_lightning as pl
 from pytorch_lightning.utilities import rank_zero_deprecation
+from pytorch_lightning.utilities.data import _replace_dataloader_init_method
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.fetching import (
     AbstractDataFetcher,
